Gtest

+++ include/ct/reflect_macros.hpp
@@ -122,19 +122,19 @@
 #define PROPERTY_WITH_FLAG_1(FLAG, NAME)                                                                               \
     constexpr static auto getPtr(const ct::Indexer<__COUNTER__ - REFLECT_COUNT_START>)                                 \
     {                                                                                                                  \
-        return ct::makeMemberPropertyPointer<FLAG>(#NAME, &DataType::NAME, nullptr);                                   \
+        return ct::makeMemberPropertyPointer<uint64_t(FLAG)>(#NAME, &DataType::NAME, nullptr);                                   \
     }
 
 #define PROPERTY_WITH_FLAG_2(FLAG, NAME, GETTER)                                                                       \
     constexpr static auto getPtr(const ct::Indexer<__COUNTER__ - REFLECT_COUNT_START>)                                 \
     {                                                                                                                  \
-        return ct::makeMemberPropertyPointer<FLAG>(#NAME, GETTER, nullptr);                                            \
+        return ct::makeMemberPropertyPointer<uint64_t(FLAG)>(#NAME, GETTER, nullptr);                                            \
     }
 
 #define PROPERTY_WITH_FLAG_3(FLAG, NAME, GETTER, SETTER)                                                               \
     constexpr static auto getPtr(const ct::Indexer<__COUNTER__ - REFLECT_COUNT_START>)                                 \
     {                                                                                                                  \
-        return ct::makeMemberPropertyPointer<FLAG>(#NAME, GETTER, SETTER);                                             \
+        return ct::makeMemberPropertyPointer<uint64_t(FLAG)>(#NAME, GETTER, SETTER);                                             \
     }
 
 #ifdef _MSC_VER
